Changes to "times" page
Moved emphasis to the what the owner is available for (as it seems to be what the users need the most in multi-skilled environments like HAS).
Introduced the term "working pattern" to see how the users understand that in this context (we know it's an expectation, but this is a temporary workaround until the product matures).
Added "lunch" as an instruction to help with understanding (based on user feedback).