Draft error for removing skills when appointment is booked
Since users will be hacking the system for demand, we will need to stop them when they try to do so while appointments are in progress.

I would prefer to use the term manage appointments rather than only having cancel as an option, but that depends on what Hawk can build by the time this goes live.